Introduce logHttp extension property to tune HTTP logging in generated logback config file, close gatling/gatling#3949
Motivation:

In all other launchers, we document a special slf4j logger to enable logging on either all or failed HTTP traffic.
The standard approach with gatling-gradle-plugin is to have it generate the logback config file, hence users are not aware of this logging option.

Modifications:

Introduce a new extension property named `logHttp` that can take the values `NONE`, `ALL` and `FAILURES` that would set the HTTP logger level respectively to the same as root, TRACE and DEBUG.

Result:

gatling-gradle-plugin users now have a convenient option for tuning HTTP logging level.