Chore: Upgrade gradle to 6.6.1
Motivation:

We want to make sure we support the latest stable release, which is 6.6.1 as of now.

We also want to be notified of deprecations so we can anticipate on the required changes in order to support the next major release (gradle 7).

Modifications:

* build the plugin with latest 6.6.1 as of now
* add 6.6.1 to the versions tested in GradleCompatibilitySpec

Result:

Plugin is now built and tested against the latest stable release.