Update GatlingPlugin to lazy load and configure tasks
Motivation
----------

When using the Gatling Gradle plugin in conjunction with Gradle Enterprise, the
build scan identified that the Gatling plugin was eagerly loading tasks into
the task graph when they did not need to be. This slows down large builds with
hundreds of modules and tasks. Since Gradle 5.0 the concept of task
configuration avoidance [1] was added. This helps speed up builds by only
registering tasks when they are needed. This commit updates the Gatling plugin
to use task configuration avoidance.

Modifications
-------------

Moving to task configuration avoidance is fairly straightforward. There is a
guide in the documentation [1]. All instances of the create() method were
changed to the corresponding register() method. Where appropriate configuration
was moved inside the Action lambda to facilitate the lazy configuration
pattern. Where tasks were being fetched by name the call site was changed to
the named() method instead. Tests were updated to correspond to these changes.

Result
------

The result is a Gradle plugin that leverages task configuration avoidance
improving build times. The plugin remains compatible with Gradle 5.0 and above.

[1] https://docs.gradle.org/current/userguide/task_configuration_avoidance.html