chore: 'Implementation-Version' in the jar plugin
Motivation:
We add plugin version in manifest when packaging a gatling enterprise simulation. For now, this value is always null because this entry is not present in the jar plugin.

More over, we send this information when user deploy a package descriptor.

Modifications:
Add a jar section in build.gradle.