fix: Fix parameters for Gradle command in publish workflow
Motivation:

In YAML, folded-style block scalars (with '>') do not insert new lines... Except when the indentation changes. So we had unexpected new lines separating the parameters from the command.

Modifications:

Use a literal-style block scalar (with '|') and add backslashes to continue the command. At least it's explicit and there is no surprising behavior.

Result:

Parameters are properly passed to the Gradle command when executing the publish workflow.